+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- cosmwasm-std: Implement `into_empty` for `QuerierWrapper`, `Deps` and
+  `DepsMut`.
+
 ## [1.3.2] - 2023-08-15
 
 ### Fixed
@@ -1790,6 +1797,7 @@ Some main points:
 
 All future Changelog entries will reference this base
 
+[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.2...HEAD
 [1.3.2]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.2.7...v1.3.0